Added delete confirmation
added titles to activities

+++ app/src/main/java/edu/pucmm/isc581/parcial2isc581/adapters/ProductsAdapter.java
@@ -1,6 +1,7 @@
 package edu.pucmm.isc581.parcial2isc581.adapters;
 
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.BitmapFactory;
 import android.provider.MediaStore;
@@ -14,6 +15,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 import edu.pucmm.isc581.parcial2isc581.R;
 import edu.pucmm.isc581.parcial2isc581.activities.CreateProductActivity;
@@ -22,6 +24,7 @@ import edu.pucmm.isc581.parcial2isc581.activities.ModifyProductActivity;
 import edu.pucmm.isc581.parcial2isc581.datamodels.ProductoModel;
 import edu.pucmm.isc581.parcial2isc581.dbModels.CategoriaDB;
 import edu.pucmm.isc581.parcial2isc581.dbModels.ProductoDB;
+import edu.pucmm.isc581.parcial2isc581.fragments.CreateUpdateProductFragment;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 
@@ -94,11 +97,19 @@ public class ProductsAdapter extends RecyclerView.Adapter<ProductsAdapter.Produc
                 context.startActivity(intent);
             });
             btnDel.setOnClickListener(v -> {
-                ProductoDB productoDB= new ProductoDB(context).open();
-                Boolean result = productoDB.delete(productoModel.getId());
-                productoDB.close();
-                if(result)
-                    adapter.deleteItem(getAdapterPosition());
+                AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(btnDel.getContext());
+                alertDialogBuilder.setTitle("Borrar producto");
+                alertDialogBuilder.setMessage("Está seguro de que quiere borrar el producto " + productoModel.getName() + "?");
+                alertDialogBuilder.setPositiveButton("Sí", (dialogInterface, i) -> {
+                    ProductoDB productoDB= new ProductoDB(context).open();
+                    Boolean result = productoDB.delete(productoModel.getId());
+                    productoDB.close();
+                    if(result)
+                        adapter.deleteItem(getAdapterPosition());
+                });
+                alertDialogBuilder.setCancelable(true);
+                alertDialogBuilder.show();
+
             });
         }
     }
